opt: v0ctemp,wctemp dims
But need more debugging for this. 
Look 
Look at the v0cwctempdim-end
TRYING TO SAVE MORE TIME BY 
CHANGING THE DIM OF V0C AND WC INSIDE THE CHOICE J LOOP IN GETDEC. MORE BELOW. 
HOw? see below. 
in sol and getdec, added vm0ctemp,vf0ctemp,wmctemp,wfctemp
The purpose wa to write them as (q,x) rather than (x,q) in order to avoid bad indexing within getdec (see how it was before this: 
In getdec, when looping over the choices j, 
I was going through the vm0_c )fpr example) 
in a bad order. 
the inner loop is over j (and so it's over i or put differently it's over q choices) but the vm0_c was indexed as vm0_c(x,q,ia,index)
AND it's doing that bad indexing within a big loop that starts with q0 adn then q and then x (this big loop is in the sol routine and then getdec is called) 
so now I rewrite those arrays with the right order indices BEFORE that big loop in sol starts.