opt issue: aws onlysingles run output
I stopped this by mistake but it had already gone quite a bit. This opt was run because the same opt was stopping in the chi cluster. So wanted to see if this would stop at aws as well and it didn't stop. also noticed that the SArands' are different between chi and aws. not sure why. the opt also follows quite a dif path but not sure if the parameters change to similar values (relative to chi cluster) for example. will try Nelder_Mead0 now.