writebest correction in obj
this was just an neval argument correction. general explanation: writebest is called from pnmead and it writes the minvalue of h and the parameters that correspond to it as well as neval,hmean,hstd. 
I also write the parameters (along with moments) that are best so far by calling from within obj. note that this is a different thing than the call of writebest from within pnmead. They should be consistent especially if the temp is low at that point (because sometimes if the temp is high then the good points might be discarded since then htherm will be quite different wrt h)