txtfil transfer: opt092822tempsmall and all other output files
This is the output of an opt that converged! But the stopcr was 10 btw. See the commit with tag opt092822tempsmall for the version of files used in this optimization. These are the outputs after I added a lot of comments to pnmead, decreased the temp parameters for annealing a lot (see main) and changed writebest functn2. SEe the tag for the exact version of fiels that were used in this optimization.