some aesthetic debugging of pnmead and objf
aesthetic because I don't think these bugs made the output of the opt0928tempsmall any wrong. The only thing was that neval was not writing properly in the call to writebest (fnctn2). So I made the dim of hmin hmin(1) instead. Also changed the names of some output files and that's it.