run2 here good! alf1t extreme vals high ptype
run2 is with more extreme values for alf1t and ptype's also different (ptype for the intermediate alft value lowered. ptypehs that is). sigwge increasing is just not going to work because it increases wdif move too much. in run2 wdif's look ok AND wvar's are ok too. mean wages seem a bit high though.
run3 was with all of above but with higher sigwge and it didn't work out because it increases wdif a lot.