normalizing alf20 for loc3 in params (and alf10)
note that the actual addition here is the norm of alf20 loc3. because alf10 loc3 was already normalized but it was not done within getpars. the step was set to 0 in getpars and the parameter itsel was also 0 but set in parameter file or main. now I made it so that alf10 loc3 is normalized within getpars.