different sigo/cst comb (m/f)+high sigwge(m/f)+high sigom (lower sigof)
This is the first run after the getting rid of cost normalization, and uncommenting out alphaed = alphaed thing in getpars adn tidying up of getmom adn adding some moments to it for joint emp trans (among others) and resetting of all weights to 1 in getpars. Trying this new version with different combos of things.