mumar/cst combos by type
to get getmar and div rates right: 
marrying types need to move sufficiently and not getdiv too much. 
maybe do a combo! increase mumar for one of the types (type 4's' mumar is right now the highest)
and then for that type make cst high in absval (relative to other types maybe or relative to sigo or something)
then that type will marry more relative to others, and div less. 
conditional on marriage, that type will be a larger proportion of the married pool. 
especially if you make the mumar of the other types lower then the married pool will consist mostly of the high mumar people and their div rates will be lower! 
yay!