another mumar/cst combos by type
made mumar for type 4 lower becuase after a certain number, going higher doesn't seem to make much of a difference (based on m100522_3). And decreased cst in absval as well (for only type4). and increased the mumar of the other types because they were not marrying at all when their mumar was low like that.