fnrphc problem end
see tag fnprhcproblem begin way before this commit.  Right before the commit with that tag, I was looking at wdif's by age and had noticed before that wdif at age 18 is too big. so I wrote wdif's by each age and then set psih(1) equal to 100 to see what happens when the prob of moving up the exp increases and eventuall is exactly 1. Then I saw that when psih increases, wdif's do increase but then when it's 1, they are all 0 and emp rates go to very very small! So then I first thought it's something wrong with the fnprof's since there were those 5's and 10's there and it seemed wrong. But it wasn't fnprof. It was fnprhc. 
migfuncobsolete: throughout this process I also noticed that I'm not even using migfunc anymore! all those functions are in mig_share! not sure when I did that , but renamed migfunc as migfuncobsolete. 
fnprhc was messing up when psih is 100 because when exp is nexp, all those probabilities were being set to 0 (when working). so then it was doing weird things and they were choosing to not work. 
corrected fnprhc. see changes. I will move some of these changes to the main branch. and will get out of this one.