wdif at age 18 too big non problem
continuing this where I left off. if you look at the fnprhcproblem-begin tag, that's the run I did when I first noticed the fnprhcproblem. I was trying to look at why wdif at age 18 is so big. So I had tried to do a run with numit=3 where I tried with psih(1) (i.e. prob exp up) very small to medium size to 1. But that messed up things when prob exp up was set to 1. see the fnprhcproblem branch below. 
This commit implements the changes I made during the debugging of that problem. 
mainly just the correction of fnprhc (see the end tag of that problem). 
there are also some aesthetic changes such as yaz. 
params psih setting is changed as well. 
and getmom is changed to write wdif by each age and nmom is changed to 1500 accordingly.