.........bug: fnmove in params was a problem
bcuz fnmove calls index2cptypehome which is a subroutine in module share. But params does not use share. IT is share that uses params. So now moved all those subroutines of transformation in share to params as well (in addition to all the funcs, which was done in prev commit)