BIG CHGE in alphaed,alphakid + yaz0 + locname
note that it is a BIG CHGE but all these were mostly aesthetic. GEtting rid of the setting of alphaed betwene ed to noed (which can be replicated by setting those parameters equal in main),getting rid of the useless dim of alphakid and writing of fnwge in yaz0 and addition of locname. objval should not have changed and indeed it has not.  in order to check that I didn't do anything wrong, can check the moment run here m100922_1 consistency with m100822_7 (the commit right before this change). 
checked and they are indeed consistent. 
note that didn't try skriv=true after I made the yaz0 aesthetic changes of writing fnwage stuff.
also note that the writing of the psih parameter in writemoments (with realpar adn temprob in getpars) is not right. but it doesn't affect anything .it's just a write mistake.