BIG CHGE in kcst in fnmove, numit=1
in fnmove we had: kcst * one(empo==np1)
have no idea why. this means that kcst was not really kid cost. it was only being applied if unemployed and had nthing to do with kids!
note that another thing I did here is an aesthetic change in getpars. There was a write mistake for psih (which was not affecting objval or anything) and I corrected this. 
numit=1 now. 
also going back to normal pars values in main (whatever normal is)