in getmom jointemp was in the rel loop
so it was being written as 0 for singles of course. just junk but harmless. moved those moments to the beginning of the moments file now. and also did some tidying up such as making the by age moments not at each age but eyvery 10 years etc. also moved around some of the by type moments. see changes. 
.....