decrease sigo's, numit=2
wdif move was negative in m100922_2. 
so it needs to incrase. recall taht one reason we had increased the sigo's was to decrease wdif mvoe because 
it used to be very large and positive
so now I think I can afford to decrease sigo's for both men and women
decrease in sigo will lead to a fall in move rates ----> this is fine since move rates are too big for both singles and couples right now. not too much though. 
decerase in sigo will lead to an increase in eu move ---> that's great! since eu move has been too low ever since we increased sigo's