opt comp for couples btw init, then best, then mumar's high, and fem eq
bcause opt was for onlysingles. And while that was running I did runs that's shown that high mumar's is the only way to bring down div move and then setting her uhome,sigo equal to his is the only way to bring up her ee move. so now will combine all of that together with the opt results from onlysingles (opt 2 was only iterating on wage related parameters to get the wages right)
numit


 onlysingles false