.......THE RUNS SINCE ADDING TAXES MIGHT NOT HAVE BEEN OK
taxbug: I did fboundscheck to figure out what was  happening in the mom file (it was just a numit and bunch misunderstanding because I forgot that when numit is 2, bunch=9 is not going to work out). but while doing this I discovered ANOTHER bug that apparently I had put in when I added taxes. sbrack=0 and pbrack=0 should have been changed when I changed the first index i.e. the first bracket from 0 to 1 but they were still being set to 0 when unemp and that was doing out of bounds. who knows what that did. but maybe nothing. anyway corrected it now.