BIG CHGE: ADDED AGECST!!! ALSO MOVESHOCKDIV WAS NOT AT RIGHT PLACE
do the same runs as previous commit to check all is ok. Also noticed that I added moveshockdiv to the wrong getdec (I added it to getdec_c). I know I am not using it anymore but these old routines might be an issue. It's risky as I might add something to the old routines instead of the new ones. Take care of  this later.