RESETTING ALFT'S
because they were so diffeerent by type, and also mar rates are different by type, the mar men wages for example were very different from that of single men and that made it very jumpy and hard to figure out psih etc. 
but now when these are reset all to 9.2, it is clear that the psih up is too high. wages increase too fast. as the up prob os around 0.2 now. which might also explain the low wdif2's. not sure though.