psih/alf12 again but different values
this whole psih/alf12 runs thing started because the last opt basically made the sigwges lower and increased psih. also slightly changed alf12 (and maybe alf22 but didn't check that one). I Tried to check why it did that. what is it benefiting. which then led me down the road of how each of those parameters are identified in the first place. and hence these runs. looked back at october 10-11's (not sure which date) where I did these runs. but it wasn't that informative. maybe it was but I still wanted to run them again. now next I will also run with sigwge changes to continue looking at this ident issue. if it's an issue at all.