sigwge/psih/alf12 runs
by the way in the last opt, the psih increased probably because there was a higher weight on wdif hmemve and it was trying to get that up. and once it got that up then it had to decrease sigwge. this was probably for bringing up wdif move in general not particularly wdif hmemove but that was the one with the higher weight on it. and it was probably because mar men's wdif mve is too low. not sure if the same thing would have happened if it was onlysingles. 
no that wasn't the reason (probably). probably what's happened is this: we had datcountbar before in the setting of msm weights, but then got rid of that after setting up stderrors. So then now it had non0 weights on wvar loc5 at 18 moments for married men for example, but there are only 5 people in that cell. Their var is low (around 0.09). So it probably was decreasing sigwge for that reason. It could still be wdif mve stuff but wdif move increasing with sigwge being a problem has been an ageold forever problem. I'm not sure if bringing down wdif move was the reason. I don't know. one of these things I think. or a combination.