psih/alf12 runs with sigwge=0.4 sth now
the higher sigwge's are too high. wvarned at loc5 pretty much nails it and that's around 0.25(?) I think. So sigwge is pinned down by that (sort of).  now I'm tyring the values of alf12 below 0.38 and psih between 0.12 and 0.27. alf12 will likely not be above 0.38 because even with 0.28, the psih's would need to be lower than 0.12 to make wages at later ages not too high. So now I'm trying psih between 0.12\s and 0.27 (see previous commmit for waht I mean).