opt 1 (wvar weights etc.)
the last opt (not prev commit), which is not recorded, ALSO took the sigwge down. I noticed that one of the things this improves is home move rate and because the weight on that moment is higher, the optimization goes there. And it doesn't care that it messes up the wvar, wdif etc. moments. So I did a lot of tidying up (such as getting rid of the nuissance weight parameters). And then I increased the weights on wages and wvar and wdif (made wdif and wvar moments the same parameter, which in retrospect I'm not sure if it iwas a good idea because wvar moments always seem to not get taht much weight whereas wdif moments do. at least currently.