trying to get wvar with alf1t instead
because it will never work out with sigw. When you look at the data you see that the wvar doesn't change that much by age. That means that wvar is mostly due to permanent heterogeneity. This is consistent with the fact that wdif gets way too high (for move especially) when we increase sigw and that does not match the data. 
.....