sending another opt with alf1t's different (higher var) and wdec wgts set to wwage
the alf1t is explained below somewhere. and the wdec weight is because it had the wvar weight (after the last changes to weights) and that I increased way way way too high in order to get the wvar's to count at all. But that was not the appropriate weight for wdecile because its contribution really shot up after that so I changed that weight to wwage.
and also by the way I stopped the last 3 opts I had started. because. well since now alf1t inits are different I guess there's no reason to keep them going. I guess I realized there's no other way to do it. Or hysteria.