some runs with divpen after multdiv changes
multdiv was only 5000 so I needed it to be higher. because there's this divmove problem and I'm not sure what exactly to do about it. I need higher divpen but then not sure if that will fix it. in any case I needed the high multdiv for this reason so there.