file transfer and compare opt 1 and 2
opt 1 ran much longer (aws stopped for opt2 but that ran quite long too). Compared to opt 1, the opt 2 had step for sigwgem 0 and ptype4 higher because in the middle of opt1 I noted that sigw was decreasing a lot adn ptype had gone down even more so started opt 2. I think opt 1 looks better probably just because it ran longer. The main issues are high move rates and very low  eumove moments for females. This is probably because the u of f is very very low now. Will try to see what happens if I increase that.