TERMINALVAL=.FALSE. now try poltax comp (poltax=0/then =6/then taxset=0)
Indeed this was it. In other words, now policytax=6 and taxset=0 give the same output. AT LAST.  So terminalval was indeed the issue. 
Note that in the poltaxcompare branch, you did the comparisons a little wrong. you first ran taxset=0 and then policytax=6, but then policytax value won't make a difference anyway since taxset is already 0 at that point. It doesn't matter anyway since the problem turned out to be terminalval, but NOTE THIS WHEN YOU ARE LOOKING AT THE OUTPUT OF THE POLTAXCOMPARE BRANCH.