now run all policytax to see what happens
See what policy effects look like after the correction of the terminalval issue. I mean it was not really fixed but I just set terminalval=.FALSE. instead so I'm just ignoring the problem rather than fixing it. Note that with termval FALSSE, getdiv move going high crazy towards end age does not happen anymore, BUT move rates increasing by age is what happens now. Which is why I had added terminalval situation. Maybe either find a better terminalval addition OR just add agecst again. I don't know. Is this really a problem. I don't know.  
What do the policytax outputs look like: It seems that only type 1 mar rates move in a different direction compared to the other types. Type 1 is the one with big big difference between alf1t and alf2t. where alf1t is much much much higher and alf2t is quite quite small. So for this type and for this type only, the mar rates seem to decrease relative to baseline. For the other types, the mar rates all INCREASE relative to the baseline. This makes sense. But then it also shows how much things are dependent on these alft's. So then I checked wdeciles whether her alf2t being so small for type 1 is justified. And it's not. These types mostly end up contributing to the mar women wages moments and the wdeciles of mar women at the lower end of the distribution is way too high. For single fems this is not as big of a problem because type 1's marry at a high rate so these type 1 women don't contribute to the single fem wage moments.