Added wcorr and wdec wgts, init alf2t and pmeet different
Because when I sent last opt, at the very beginning it decreased her emp cur f A LOT which increased mar rates AND decreased mar men wages even at age 18 quite a bit. So pmeet init is to get the mar rates more correct so it doesn't do things like that. The increase in alf2t is to get the corr moments more correctly together with wdecile.