changed some steps,and also init pars
the pars that I'm playing with here are mostly offer probs as well as alft's. The steps were changed because this was sent as an opt but it didn't run for long. So just forget about that and try some pars on your own as per usual. But note that that opt did do one thing and that was increasing the alf1t type 4 parameter. In the short amount of time the opt ran, that is. 
Also note that mar fem wages are ok at the beginning but they just don't increase. since the single fem's wages increase ok, this must be something about the married fem not pursuing job opportunities. How do I get those mar fem wages to increase by age properly?