alphaed decrease more, also offer probs
alphaed was too high so the previous pars bump didn't really work out. Also the offer probs in previous runs were changed so that his u of m was much smaller. But this is np good because then his eu move rates are too low (goes from 0.8 to 0.4 something). I had decreased in order to bring eu move down (it's 0.65 in data) but overshot it. I thought Maybe a lower value of u of m could fix this but then eu move decerased but too much.