change the order of fem=male pars blocks to see better
I set fem and male pars equal block by block but the first block was offerprobs. And that basically decreased the mar rates A LOT. So it was difficut to see which block is doing what to the wages. So now I changed the order and put sigw first (as that was the one that seemed to have a big impact on wages but it's hard to say due to the low mar rates in that run (which is the previous commit)).