now increase her alpha noed to get those emp rates down
result: when her alphaed (for ned) is higher, her wages are indeed much higher now. AND they increase more with age (though I have to verify this). The main issue is with ee move being too low now. But I will try to increase it with offer prob pars. I think maybe that's why the iteration kept not increasing this alpha (because it messed up the ee move trans rates). But then it can be fixed with offer probs. We shall see. Also need to check what happened to single fems wages by the way.