try lower empcurm (her alphaed seems fine now possibly and she has good wdif too)
so why are mar fem and single fem and single men having positive wdif but not mar men? Try lower emp curm for him which is maybe why she has good wdif too. but then I don't know because that's the case for single men as well. so this is probably doomed from the getgo.