should I get her eu stay and eu move rates lower alpha or with lower probs?
I think the latter works better. Because recall that for singles some of those trans rats were ok. And changing alpha messes with both single fem AND mar men. Or at least I think so. Now the other thing is that wdif for mar men is good and positive but mar women is positive also. as well as some of the singles I think. what to do. note that conditional one hmemove it's ok. so maybe home move rates are not high enouhg. I know they are not high enough anyway. Also the wdif's in later ages for mar fem ARE negative. it's just that there's not enough of them I suppose.