now her sigwage higher and then increase mumar and also pmeet
note that here getdiv stay starts being higher (it's been close to 0 for a while). what does it? the culprit seems to be sigwge. in _15 I increased sigwgem (and decreased cst) and the getdiv stay rates increased slightly. here I increased sigwgef as well and it seems to have REALLY increase getdiv stay rates this time