sending an opt
but this one did not really work out well. it brought mar rates down a lot. her emp rates are too low as well. it did this because it increase alphaed (as per usual). I checked why it insists on this and it seems alphaed higher really fixes the getdiv move rates. and there is higher weights on those rel moments. anyway, I'll probably use the alft pars from this opt but not the rest maybe. I don't know. it did run quite a bit. so that's not the issue I think.