opt compare (in blocks)
why did alpha(ned f) decrease so much? well when it decreases, that decreases the wned accepted wages a lot. and because they were too high and they have a high weight on them, the opt liked that even though it messed up the emp ned f rates (way too high as a result of lower alpha). because weights on emp not that high, the opt was ok with this sacrifice. so it's not necessarily the getdiv move moments that was the culprit here for this decrease in alpha(ned f). it might have something to do with it too though.