uloc=0,cst=0,uhomet=25K
I think that with the higher uhome's the moving rates become increasing by age more. Which sort of makes sense? Since they keep moving back home? I don't know. And also higher uhome is a problem because it makes wdif hmemve's too large (large positive for =1 and large negative for =0). but note that this does depend on the psil and sigo.