o121522_1 termval true
here wdif's are small also because sigwage is small. in the next run when sigwgage is higher, wdif move's shoot up again. the reason why I looked at o121522_1 pars run was because I noticed that wdif move's used to be much smaller. But it turns out that was mostly due to the low sigwage. In general, to get wdif move smaller you either need low sigwage OR high sigo. So you definitely need to increase sigo if you want higher sigwage (to get the wvar's high enough while not shooting up wdif move). Indeed in the next run, I increase sigwage and the wdif move's shoot up. I need to have higher sigo. (this is a retrospective note so don't get confused)