BIG CHGE: ADDED AGECST AGAIN!!
also deleted sig_uloc in params which was a nuissance parameter. The other nuissance parameters are nowo alf13 and alf23 (alf23 was assigned uhomet before and alf13 is now agecst). But noticed that alf13 and alf23 are in fnwge still but they are NOT initiated. So printing their values just to see if they are indeed zero. OnceI make sure, then I will delete them for good or initiate them to zero. 
....