BIG CHGE: ADDED UHOMEIA
still have agecst too. put the new agehome parameter to where uhome(1) used to be (pars22). and moved the agecst parameter to where alf13 used to be to where uhome(2) used to be. ran this check to compare to m122822_4 and it seems fine (agehome and agecst=0 give the same objval's as last run)