sigom/sigof/uloc/agehome/psil higher cst/agehome k*500/lower sigwge/higher alf12 with higher uhomet
this is the same as _20 but with all sigo loops (from the smallest to largest) AND uhomet=5000. see if uhomt makes initial moves smaller. and also want all the sigo runs in one place.
IGNORE this. forgot to change numit here. but also the later move rates were too low (so the high uhomet it doesn't work).