sigom/sigof/uloc/agehome/psil higher cst/agehome k*500/lower sigwge/higher alf12 with uhomet=30K
this is the same as _20 but with all sigo loops (from the smallest to largest). aftet this, also see see if uhomt makes initial moves smaller. and also want all the sigo runs in one place.
note that the previous one was similar but ignore it beause numit was not changed. 
IGNORE THIS ONE TOO. I CHANGED MY MIND. HIGHER UHOMET WILL NOT WORK.